Add a tool to get download stats (#33)
* make issue-stats.py mode 755 again

* Add a first cut tool to get download stats from github

Right now it only downloads the download stats for selected repos
to a date stamped file. This is useful for exploring what the
data looks like.

What we really need is to delta this against yesterday's file
to get the download count per day. That will require a permanent
store.

* better .gitignore

* fix 403 on repos we do not own