delta.py - tool to compute and update download counts by day.
If looks for downloads < 0 and the last day of data was the
previous day, just compute the download delta and use that.

If there was a gap of several days in the data, use the average
over that span. We do this even if the download count contains
a value. That corrects for some past data smoothing.