Rollup changes to download stats gathering (#42)
* delta.py - tool to compute and update download counts by day.

If looks for downloads < 0 and the last day of data was the
previous day, just compute the download delta and use that.

If there was a gap of several days in the data, use the average
over that span. We do this even if the download count contains
a value. That corrects for some past data smoothing.

* More repos and better categorization.

Major rewrite to categorization engine.
Fix python style.

* Fix for patterns like buildozer.osx

* remove shortened list of repos so we get them all

* give better error on repos like skydoc which do not have their assets in the right place

* fix file comment